ex9

+++ Atividades/Atividade 2/ex9.c
@@ -13,9 +13,9 @@ int main() {
     printf("N3: ");
     scanf("%f", &n3);
 
-    final_avarage = (2 * n1 + 3 * n2 + 5 * n3) / 10;
+    final_avarage = (2 * n1 + 3 * n2 + 5 * n3) / (float)10;
 
-    printf("Média final: %.1f\n", final_avarage);
+    printf("Média final: %.2f\n", final_avarage);
 
     return 0;
 }
